Fix: Error when setting the chosen mirror variable $CHSENMIR.
The Mirror is chosen using the
`gsm.operator.iso-country`\`gsm.sim.operator.iso-country` android
property. However, when the string started with a coma (probably caused
by having a sim card only in the sim2 slot?), the $USERCOUNTRYCODE
variable ended up being empty. This is fixed by selecting the string
after the coma when the reported iso-country starts with a coma.
Otherwise it selects what comes before the coma.